Developer

+++ src/application/services/UserService.ts
@@ -1,15 +1,15 @@
 import { inject, injectable } from 'tsyringe';
 import { IUserRepository } from '../../domain/repositories/IUserRepository';
-import { UserValidationService } from '../validation/UserValidationService';
+import { UserValidationService } from '../../domain/validation/UserValidationService';
 import { User } from '../../domain/entities/User';
 import { CreateUserDto } from '../dtos/user/CreateUserDto';
 import { UpdateUserDto } from '../dtos/user/UpdateUserDto';
 import { ResponseUserDto } from '../dtos/user/ResponseUserDto';
 import { UserMapper } from '../../application/mappers/UserMappers';
-import bcrypt from 'bcrypt';
+import bcrypt from 'bcryptjs';
 import { UserFilter } from '../../domain/interface/UserFilter';
 import { UserRole } from '../../domain/enums/UserRole';
-import { PasswordHasher } from '../../application/utils/PasswordHasher';
+import { PasswordHasher } from '../../infrastructure/shared/utils/PasswordHasher';
 
 @injectable()
 export class UserService {
@@ -46,12 +46,23 @@ export class UserService {
     return this.userMapper.toResponseUserDto(updatedUser);
   }
 
-  public async deleteUser(id: string, loggedInUserId: string): Promise<void> {
-    await this.userValidationService.validateUserExistence(id);
+  public async deleteUser(
+    id: string,
+    loggedInUserId: string,
+    providedDeleteKey?: string,
+  ): Promise<void> {
+    const user = await this.userValidationService.validateUserExistence(id);
+
     await this.userValidationService.validateDeleteSelfOperation(
       id,
       loggedInUserId,
     );
+
+    await this.userValidationService.validateDeleteKeyForDefaultAdmin(
+      user,
+      providedDeleteKey,
+    );
+
     await this.userRepository.remove(id);
   }
